requirements.txt
streamlit>=1.25.0
google-generativeai>=0.3.1
gtts>=2.3.2
PyMuPDF>=1.22.5
requests>=2.31.0
python-dotenv>=1.0.0
arxiv>=1.4.5
